fix(pagination): Fixes an issue with aria label name clashes for a11y.
Specific inputs like `aria-label-next` and `aria-label-current` are not allowed based on
the aria specification. Only known aria attributes are allowed to be
used. We have renamed the inputs in question to their camelcased
version, but kept the old ones for compatibility for now.
With version 7.0. we sould remove the kebap-cased inputs.

The role `navigation` is not allowed on ol/ul elements as it would
highjack the natural accessibility of the list. We have removed the
navigation role from the pagination root list and added it to the host
element.

Fixes #79